# Prevent `iv` parameter leaking, force BCE using full slice expressions 