--- a/encrypt.go
+++ b/encrypt.go
@@ -48,11 +48,11 @@ func EncryptBlocks(block cipher.Block, iv, dst, src []byte) {
 	m := iv[b:]
 
 	for o := 0; o < len(src); o += b {
-		xor.Bytes(dst[o:o+b], src[o:o+b], c)
-		block.Encrypt(dst[o:o+b], dst[o:o+b])
-		xor.Bytes(dst[o:o+b], dst[o:o+b], m)
+		xor.Bytes(dst[o:o+b:o+b], src[o:o+b:o+b], c)
+		block.Encrypt(dst[o:o+b:o+b], dst[o:o+b:o+b])
+		xor.Bytes(dst[o:o+b:o+b], dst[o:o+b:o+b], m)
 
-		c = dst[o : o+b]
-		m = src[o : o+b]
+		c = dst[o : o+b : o+b]
+		m = src[o : o+b : o+b]
 	}
 }