gotd `Write` benchmark benchstat:
```
name              old time/op    new time/op    delta
Write/16b-12        1.83µs ± 1%    1.76µs ± 2%   -3.38%  (p=0.008 n=5+5)
Write/128b-12       2.34µs ± 1%    2.28µs ± 0%   -2.63%  (p=0.008 n=5+5)
Write/1024b-12      5.92µs ± 1%    5.75µs ± 1%   -2.83%  (p=0.008 n=5+5)
Write/8192b-12      34.3µs ± 1%    33.1µs ± 0%   -3.71%  (p=0.008 n=5+5)
Write/65536b-12      255µs ± 1%     249µs ± 1%   -2.53%  (p=0.008 n=5+5)
Write/524288b-12    1.98ms ± 1%    1.93ms ± 1%   -2.45%  (p=0.008 n=5+5)

name              old speed      new speed      delta
Write/16b-12      8.76MB/s ± 1%  9.07MB/s ± 2%   +3.52%  (p=0.008 n=5+5)
Write/128b-12     54.8MB/s ± 1%  56.3MB/s ± 0%   +2.69%  (p=0.008 n=5+5)
Write/1024b-12     173MB/s ± 1%   178MB/s ± 1%   +2.92%  (p=0.008 n=5+5)
Write/8192b-12     239MB/s ± 1%   248MB/s ± 0%   +3.85%  (p=0.008 n=5+5)
Write/65536b-12    257MB/s ± 1%   264MB/s ± 1%   +2.60%  (p=0.008 n=5+5)
Write/524288b-12   265MB/s ± 1%   271MB/s ± 1%   +2.52%  (p=0.008 n=5+5)

name              old alloc/op   new alloc/op   delta
Write/16b-12          752B ± 0%      720B ± 0%   -4.26%  (p=0.008 n=5+5)
Write/128b-12         864B ± 0%      832B ± 0%   -3.70%  (p=0.008 n=5+5)
Write/1024b-12      1.82kB ± 0%    1.79kB ± 0%   -1.75%  (p=0.008 n=5+5)
Write/8192b-12      10.2kB ± 0%    10.1kB ± 0%   -0.32%  (p=0.008 n=5+5)
Write/65536b-12     74.5kB ± 0%    74.5kB ± 0%   -0.04%  (p=0.016 n=5+5)
Write/524288b-12     534kB ± 0%     534kB ± 0%     ~     (p=0.730 n=5+4)

name              old allocs/op  new allocs/op  delta
Write/16b-12          7.00 ± 0%      6.00 ± 0%  -14.29%  (p=0.008 n=5+5)
Write/128b-12         7.00 ± 0%      6.00 ± 0%  -14.29%  (p=0.008 n=5+5)
Write/1024b-12        7.00 ± 0%      6.00 ± 0%  -14.29%  (p=0.008 n=5+5)
Write/8192b-12        7.00 ± 0%      6.00 ± 0%  -14.29%  (p=0.008 n=5+5)
Write/65536b-12       7.00 ± 0%      6.00 ± 0%  -14.29%  (p=0.008 n=5+5)
Write/524288b-12      7.00 ± 0%      6.00 ± 0%  -14.29%  (p=0.008 n=5+5)
```

BCE improvment benchstat: 
```
name                     old time/op    new time/op    delta
DecryptBlocks/16-12         256ns ± 1%     254ns ± 2%    ~     (p=0.333 n=5+5)
DecryptBlocks/128-12        413ns ± 1%     402ns ± 1%  -2.61%  (p=0.008 n=5+5)
DecryptBlocks/1024-12      1.56µs ± 0%    1.50µs ± 0%  -3.57%  (p=0.008 n=5+5)
DecryptBlocks/8192-12      10.8µs ± 1%    10.2µs ± 1%  -5.48%  (p=0.008 n=5+5)
DecryptBlocks/65536-12     85.0µs ± 1%    80.2µs ± 0%  -5.68%  (p=0.008 n=5+5)
DecryptBlocks/524288-12     678µs ± 1%     640µs ± 0%  -5.57%  (p=0.008 n=5+5)
EncryptBlocks/16-12         261ns ± 1%     249ns ± 2%  -4.64%  (p=0.008 n=5+5)
EncryptBlocks/128-12        416ns ± 0%     401ns ± 0%  -3.71%  (p=0.008 n=5+5)
EncryptBlocks/1024-12      1.58µs ± 1%    1.51µs ± 1%  -4.78%  (p=0.008 n=5+5)
EncryptBlocks/8192-12      11.1µs ± 1%    10.3µs ± 0%  -6.96%  (p=0.008 n=5+5)
EncryptBlocks/65536-12     86.0µs ± 1%    80.8µs ± 1%  -6.04%  (p=0.008 n=5+5)
EncryptBlocks/524288-12     685µs ± 1%     645µs ± 0%  -5.73%  (p=0.008 n=5+5)

name                     old speed      new speed      delta
DecryptBlocks/16-12      62.5MB/s ± 1%  63.1MB/s ± 2%    ~     (p=0.333 n=5+5)
DecryptBlocks/128-12      310MB/s ± 1%   318MB/s ± 1%  +2.68%  (p=0.008 n=5+5)
DecryptBlocks/1024-12     658MB/s ± 0%   683MB/s ± 0%  +3.71%  (p=0.008 n=5+5)
DecryptBlocks/8192-12     757MB/s ± 1%   801MB/s ± 1%  +5.79%  (p=0.008 n=5+5)
DecryptBlocks/65536-12    771MB/s ± 1%   817MB/s ± 0%  +6.02%  (p=0.008 n=5+5)
DecryptBlocks/524288-12   773MB/s ± 1%   819MB/s ± 0%  +5.90%  (p=0.008 n=5+5)
EncryptBlocks/16-12      61.4MB/s ± 1%  64.4MB/s ± 2%  +4.87%  (p=0.008 n=5+5)
EncryptBlocks/128-12      308MB/s ± 0%   319MB/s ± 0%  +3.85%  (p=0.008 n=5+5)
EncryptBlocks/1024-12     647MB/s ± 1%   680MB/s ± 1%  +5.01%  (p=0.008 n=5+5)
EncryptBlocks/8192-12     738MB/s ± 1%   793MB/s ± 0%  +7.47%  (p=0.008 n=5+5)
EncryptBlocks/65536-12    762MB/s ± 1%   811MB/s ± 1%  +6.43%  (p=0.008 n=5+5)
EncryptBlocks/524288-12   766MB/s ± 1%   812MB/s ± 0%  +6.08%  (p=0.008 n=5+5)

name                     old alloc/op   new alloc/op   delta
DecryptBlocks/16-12          448B ± 0%      448B ± 0%    ~     (all equal)
DecryptBlocks/128-12         448B ± 0%      448B ± 0%    ~     (all equal)
DecryptBlocks/1024-12        448B ± 0%      448B ± 0%    ~     (all equal)
DecryptBlocks/8192-12        448B ± 0%      448B ± 0%    ~     (all equal)
DecryptBlocks/65536-12       448B ± 0%      448B ± 0%    ~     (all equal)
DecryptBlocks/524288-12      448B ± 0%      448B ± 0%    ~     (all equal)
EncryptBlocks/16-12          448B ± 0%      448B ± 0%    ~     (all equal)
EncryptBlocks/128-12         448B ± 0%      448B ± 0%    ~     (all equal)
EncryptBlocks/1024-12        448B ± 0%      448B ± 0%    ~     (all equal)
EncryptBlocks/8192-12        448B ± 0%      448B ± 0%    ~     (all equal)
EncryptBlocks/65536-12       448B ± 0%      448B ± 0%    ~     (all equal)
EncryptBlocks/524288-12      448B ± 0%      448B ± 0%    ~     (all equal)

name                     old allocs/op  new allocs/op  delta
DecryptBlocks/16-12          4.00 ± 0%      4.00 ± 0%    ~     (all equal)
DecryptBlocks/128-12         4.00 ± 0%      4.00 ± 0%    ~     (all equal)
DecryptBlocks/1024-12        4.00 ± 0%      4.00 ± 0%    ~     (all equal)
DecryptBlocks/8192-12        4.00 ± 0%      4.00 ± 0%    ~     (all equal)
DecryptBlocks/65536-12       4.00 ± 0%      4.00 ± 0%    ~     (all equal)
DecryptBlocks/524288-12      4.00 ± 0%      4.00 ± 0%    ~     (all equal)
EncryptBlocks/16-12          4.00 ± 0%      4.00 ± 0%    ~     (all equal)
EncryptBlocks/128-12         4.00 ± 0%      4.00 ± 0%    ~     (all equal)
EncryptBlocks/1024-12        4.00 ± 0%      4.00 ± 0%    ~     (all equal)
EncryptBlocks/8192-12        4.00 ± 0%      4.00 ± 0%    ~     (all equal)
EncryptBlocks/65536-12       4.00 ± 0%      4.00 ± 0%    ~     (all equal)
EncryptBlocks/524288-12      4.00 ± 0%      4.00 ± 0%    ~     (all equal)
```